+++ src/idpconfgen/components/bgeo_strategies/sampling/__init__.py
@@ -1,0 +1,11 @@
+"""
+Sample bond geometries from a database.
+
+This database was created from high-resolution PDB structures.
+"""
+from pathlib import Path
+
+
+_file = Path(__file__).parent
+bgeo_sampling_path = Path(_file, 'bgeo_sampling.tar')
+name = "sampling"