+++ CHANGELOG.rst
@@ -9,7 +9,13 @@ While in version ``0``, minor and patch upgrades converge in the ``patch`` numbe
 Changelog
 =========
 
-* updated GRAHAM install instructions
+* updated GRAHAM install instruction (#207)
+
+v0.6.0 (2022-05-24)
+------------------------------------------------------------
+
+* Add bond geometry option to build with `Int2Cart` software
+* PR #203
 
 v0.5.1 (2022-05-24)
 ------------------------------------------------------------